# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8089","description":"Generated server url"}],"paths":{"/api/projects/{projectId}/tasks/{id}":{"get":{"tags":["task-controller"],"operationId":"getTaskById","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["task-controller"],"operationId":"updateTask","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["task-controller"],"operationId":"deleteTask","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/projects/{projectId}/tasks/{id}/move":{"put":{"tags":["task-controller"],"operationId":"moveTask","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":true,"schema":{"type":"string","enum":["TODO","IN_PROGRESS","DONE"]}},{"name":"position","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/projects/{projectId}/labels/{id}":{"get":{"tags":["label-controller"],"operationId":"getLabelById","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["label-controller"],"operationId":"updateLabel","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LabelRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["label-controller"],"operationId":"deleteLabel","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/projects/{id}":{"get":{"tags":["project-controller"],"operationId":"getProjectById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["project-controller"],"operationId":"updateProject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["project-controller"],"operationId":"deleteProject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/projects":{"get":{"tags":["project-controller"],"operationId":"getAllProjects","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"}}}}}}},"post":{"tags":["project-controller"],"operationId":"createProject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/projects/{projectId}/tasks":{"get":{"tags":["task-controller"],"operationId":"getAllTasks","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["task-controller"],"operationId":"createTask","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/projects/{projectId}/labels":{"get":{"tags":["label-controller"],"operationId":"getAllLabels","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["label-controller"],"operationId":"createLabel","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LabelRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/projects/{id}/members/{userId}":{"post":{"tags":["project-controller"],"operationId":"addMember","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["project-controller"],"operationId":"removeMember","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/signout":{"post":{"tags":["auth-controller"],"operationId":"logoutUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/signin":{"post":{"tags":["auth-controller"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/refreshtoken":{"post":{"tags":["auth-controller"],"operationId":"refreshtoken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenRefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/api/users/search":{"get":{"tags":["user-controller"],"operationId":"searchUsers","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/api/users/me":{"get":{"tags":["user-controller"],"operationId":"getCurrentUserProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/projects/{projectId}/tasks/status/{status}":{"get":{"tags":["task-controller"],"operationId":"getTasksByStatus","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"path","required":true,"schema":{"type":"string","enum":["TODO","IN_PROGRESS","DONE"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/projects/{projectId}/tasks/search":{"get":{"tags":["task-controller"],"operationId":"searchTasks","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/projects/search":{"get":{"tags":["project-controller"],"operationId":"searchProjects","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"}}}}}}}}},"components":{"schemas":{"TaskRequest":{"required":["title"],"type":"object","properties":{"title":{"maxLength":100,"minLength":0,"type":"string"},"description":{"maxLength":1000,"minLength":0,"type":"string"},"status":{"type":"string","enum":["TODO","IN_PROGRESS","DONE"]},"priority":{"type":"string","enum":["LOW","MEDIUM","HIGH"]},"position":{"type":"integer","format":"int32"},"dueDate":{"type":"string","format":"date-time"},"assigneeId":{"type":"integer","format":"int64"},"labelIds":{"uniqueItems":true,"type":"array","items":{"type":"integer","format":"int64"}}}},"LabelRequest":{"required":["name"],"type":"object","properties":{"name":{"maxLength":50,"minLength":0,"type":"string"},"color":{"maxLength":7,"minLength":0,"type":"string"}}},"ProjectRequest":{"required":["name"],"type":"object","properties":{"name":{"maxLength":100,"minLength":0,"type":"string"},"description":{"maxLength":500,"minLength":0,"type":"string"},"status":{"type":"string","enum":["ACTIVE","ARCHIVED","COMPLETED"]}}},"SignupRequest":{"required":["email","firstName","lastName","password","username"],"type":"object","properties":{"username":{"maxLength":50,"minLength":3,"type":"string"},"email":{"maxLength":100,"minLength":0,"type":"string"},"firstName":{"maxLength":100,"minLength":0,"type":"string"},"lastName":{"maxLength":100,"minLength":0,"type":"string"},"password":{"maxLength":40,"minLength":6,"type":"string"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"TokenRefreshRequest":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string"}}},"User":{"required":["email","firstName","lastName","username"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"maxLength":50,"minLength":0,"type":"string"},"email":{"maxLength":100,"minLength":0,"type":"string"},"firstName":{"maxLength":100,"minLength":0,"type":"string"},"lastName":{"maxLength":100,"minLength":0,"type":"string"},"role":{"type":"string","enum":["USER","ADMIN","PROJECT_MANAGER"]},"enabled":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"Project":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":100,"minLength":0,"type":"string"},"description":{"maxLength":500,"minLength":0,"type":"string"},"status":{"type":"string","enum":["ACTIVE","ARCHIVED","COMPLETED"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"owner":{"$ref":"#/components/schemas/User"},"members":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}